feat: save failed CSV rows to S3
